Updates & Left/Right side support
Updated GradleRIO and added support for allowing the close command to work for both the left and right sides by making the angle negative if the side is the right (assuming that the commands are written for the left side).

May need to make changes/move side multiplier calculation to a method, will change as necessary.